Code cleanup

+++ loan_equipment_manager/static/loan_equipment_manager/main.js
@@ -14,6 +14,10 @@ let alertBox;
 window.addEventListener('DOMContentLoaded', (event) => {
     console.log(location.pathname);
 
+    //Sorts items table when page accessed     
+    //Tables.sortItemsTable("itemsTable");  
+    Tables.sortAllTables();  
+
     if(location.pathname === "/loan_request"){
 
         //Reference to request form
@@ -29,46 +33,38 @@ window.addEventListener('DOMContentLoaded', (event) => {
         //Sets min date for date picker
         Helpers.setMinLoanDate();
     }
+
+    if(location.pathname === "/manage_requests"){
+        //Add event listeners to all Accept buttons in requests page
+        //Set request as accepted on submit.
+        Requests.reqAccept();
+
+        //Add event listeners to all Reject buttons in requests page
+        //Set request as rejected on submit.
+        Requests.reqReject();
+    }
+
+    if(location.pathname === "/manage_equipment"){
+        //Adds event listener to Save and Update item button
+        //Takes form data and submits to the server
+        Equipment.editLoanItem();
+
+        //Adds event listener to Add loan equipment button
+        //Submits form data to the server
+        Equipment.addLoanItem();  
+
+        //Adds event listener to Delete item button
+        //Submits a request to the server to delete item
+        Equipment.deleteItem(); 
+    }
+
+    if(location.pathname === "/manage_loans"){
+        //Adds event listener to Close item buttons
+        //Sends a request to the server to mark loan as closed
+        Loans.closeLoan();
+    }
     
-    
 
-    
-
-    
-
-    //Sorts items table when page accessed     
-    //Tables.sortItemsTable("itemsTable");  
-    Tables.sortAllTables();  
-
-    //Adds event listener to Update Item buttons when found
-    //Populates all fields with data from current row
-    Tables.populateModalWindow();   
-
-
-
-    //Add event listeners to all Accept buttons in requests page
-    //Set request as accepted on submit.
-    Requests.reqAccept();
-
-    //Add event listeners to all Reject buttons in requests page
-    //Set request as rejected on submit.
-    Requests.reqReject();
-
-    //Adds event listener to Save and Update item button
-    //Takes form data and submits to the server
-    Equipment.editLoanItem();
-
-    //Adds event listener to Add loan equipment button
-    //Submits form data to the server
-    Equipment.addLoanItem();  
-
-    //Adds event listener to Delete item button
-    //Submits a request to the server to delete item
-    Equipment.deleteItem();  
-
-    //Adds event listener to Close item buttons
-    //Sends a request to the server to mark loan as closed
-    Loans.closeLoan();
